requirements.txt
fastmcp>=3.0.0
fastapi>=0.100.0
psycopg2-binary>=2.9.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
